# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A control plane for Chorus Worker that manages S3 replication jobs",
        "title": "Chorus Controller API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/replications": {
            "get": {
                "description": "Returns a list of all configured replication jobs with their statuses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replications"
                ],
                "summary": "List all replication jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Configures a new replication job for specified buckets between storages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replications"
                ],
                "summary": "Create a new replication job",
                "parameters": [
                    {
                        "description": "Replication configuration",
                        "name": "replication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateReplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Replication job created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a replication job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replications"
                ],
                "summary": "Delete a replication job",
                "parameters": [
                    {
                        "description": "Replication identifier",
                        "name": "replication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReplicationIdentifier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Replication deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/replications/pause": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pauses an active replication job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replications"
                ],
                "summary": "Pause a replication job",
                "parameters": [
                    {
                        "description": "Replication identifier",
                        "name": "replication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReplicationIdentifier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Replication paused successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/replications/resume": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resumes a paused replication job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replications"
                ],
                "summary": "Resume a paused replication job",
                "parameters": [
                    {
                        "description": "Replication identifier",
                        "name": "replication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReplicationIdentifier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Replication resumed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/replications/switch/zero-downtime": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Switches main and follower buckets for a replication job without blocking writes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replications"
                ],
                "summary": "Switch main and follower buckets without downtime",
                "parameters": [
                    {
                        "description": "Replication identifier",
                        "name": "replication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReplicationIdentifier"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Switch initiated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.CreateReplicationRequest": {
            "type": "object",
            "required": [
                "from",
                "to",
                "user"
            ],
            "properties": {
                "agent_url": {
                    "type": "string"
                },
                "buckets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "to_bucket": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "domain.ReplicationIdentifier": {
            "type": "object",
            "required": [
                "bucket",
                "from",
                "to",
                "user"
            ],
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "to_bucket": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TokenAuth": {
            "description": "Type \"Token\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}